make $8000000 a valid long integer (-2147483648)

+++ intermediate/src/prog8/intermediate/IRFileWriter.kt
@@ -272,7 +272,7 @@ class IRFileWriter(private val irProgram: IRProgram, outfileOverride: Path?) {
                 dt.isBool -> constant.value.toInt().toString()
                 dt.isFloat -> constant.value.toString()
                 dt.isPointer -> TODO("constant pointer $constant")
-                dt.isInteger -> constant.value.toInt().toHex()
+                dt.isInteger -> constant.value.toLong().toHex()
                 else -> throw InternalCompilerException("weird dt")
             }
             xml.writeCharacters("${constant.typeString} ${constant.name}=$value\n")
